Like, seznam výrobků z podledu majitele

+++ wp-content/plugins/budkutil/js/like.php
@@ -34,34 +34,50 @@ if(is_user_logged_in())
 	$produkt_id = $_GET['pid'];
 	$user_id 	= get_current_user_id();
 	
+	$args = array(
+    	'ID'        =>  $produkt_id, 
+    );
+	$pos = get_posts( $args );
 
-	$posledni = $wpdb->get_var("SELECT time FROM bk_like WHERE user_id='".$user_id."' ORDER BY time DESC");
-
-	if($posledni < time()-2)
+	if($post->post_author != $user_id)
 	{
-		if(is_null($wpdb->get_var("SELECT time FROM bk_like WHERE user_id='".$user_id."' AND produkt_id='".$produkt_id."'")))
-		{
-			$wpdb->insert('bk_like', 
-		        array(  
-		            'user_id' => $user_id, 
-		            'produkt_id' => $produkt_id,
-		            'time' => time(),
-		            'ip' => get_ip()
-		        )
-		    );
 
-		    exit_status('liked');
+		$posledni = $wpdb->get_var("SELECT time FROM bk_like WHERE user_id='".$user_id."' ORDER BY time DESC");
+
+		if($posledni < time()-2)
+		{
+			if(is_null($wpdb->get_var("SELECT time FROM bk_like WHERE user_id='".$user_id."' AND produkt_id='".$produkt_id."'")))
+			{
+				$wpdb->insert('bk_like', 
+			        array(  
+			            'user_id' 		=> $user_id, 
+			            'produkt_id' 	=> $produkt_id,
+			            'time' 			=> time(),
+			            'ip' 			=> get_ip()
+			        )
+			    );
+
+			    exit_status('liked');
+			}
+			else
+			{
+				$wpdb->delete( 'bk_like', array( 'produkt_id' => $produkt_id, 'user_id' => $user_id  ) );
+
+				exit_status('annulled');
+			}
 		}
 		else
 		{
-			$wpdb->delete( 'bk_like', array( 'produkt_id' => $produkt_id, 'user_id' => $user_id  ) );
-
-			exit_status('annulled');
+			exit_status('Small time interval');
 		}
 	}
 	else
 	{
-		exit_status('Small time interval');
+		exit_status('Author');
 	}
+}
+else
+{
+	exit_status('Loged');
 }
 ?>
